+++ packages/ubcli/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [5.1.2]
+### Fixed 
+ - [unitybase/ubjs#15] - Postgre DDl generator must use `SELECT nextval('${sequenceObj}')` for sequence incrementing
+
 ## [5.1.0]
 ### Added
 - `ubcli generateDDL` now support Json type attributes